Move ImageBlock to its own app
